Develop smc to master SMC

+++ apps/health_campaign_field_worker_app/lib/blocs/auth/auth.dart
@@ -3,13 +3,18 @@ import 'package:digit_components/digit_components.dart';
 import 'package:dio/dio.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
+import '../../data/data_repository.dart';
 import '../../data/local_store/secure_store/secure_store.dart';
 import '../../data/repositories/remote/auth.dart';
 import '../../data/repositories/remote/mdms.dart';
 import '../../models/auth/auth_model.dart';
+import '../../models/entities/boundary.dart';
 import '../../models/role_actions/role_actions_model.dart';
+import '../../utils/background_service.dart';
 import '../../utils/environment_config.dart';
 import '../../utils/utils.dart';
+import '../../models/data_model.dart';
+import '../../utils/constants.dart';
 
 // part 'auth.freezed.dart' need to be added to auto generate the files for freezed model
 part 'auth.freezed.dart';
@@ -21,9 +26,12 @@ class AuthBloc extends Bloc<AuthEvent, AuthState> {
   final LocalSecureStore localSecureStore;
   final AuthRepository authRepository;
   final MdmsRepository mdmsRepository;
+  final LocalRepository<BoundaryModel, BoundarySearchModel>
+      boundaryLocalRepository;
 
   AuthBloc({
     required this.authRepository,
+    required this.boundaryLocalRepository,
     required this.mdmsRepository,
     LocalSecureStore? localSecureStore,
   })  : localSecureStore = LocalSecureStore.instance,
@@ -31,6 +39,7 @@ class AuthBloc extends Bloc<AuthEvent, AuthState> {
     on(_onLogin);
     on(_onLogout);
     on(_onAutoLogin);
+    on(_onAuthLogoutWithoutToken);
   }
 
   //_onAutoLogin event handles auto-login of the user when the user is already logged in and token is not expired, AuthenticatedWrapper is returned in UI
@@ -80,16 +89,19 @@ class AuthBloc extends Bloc<AuthEvent, AuthState> {
       await localSecureStore.setAuthCredentials(result);
       await localSecureStore.setBoundaryRefetch(true);
 
-      final actionsWrapper = await mdmsRepository
-          .searchRoleActions(envConfig.variables.actionMapApiPath, {
-        "roleCodes": result.userRequestModel.roles.map((e) => e.code).toList(),
-        "tenantId": envConfig.variables.tenantId,
-        "actionMaster": "actions-test",
-        "enabled": true,
-      });
+      final actionsWrapper = await mdmsRepository.searchRoleActions(
+        envConfig.variables.actionMapApiPath,
+        localSecureStore,
+        {
+          "roleCodes":
+              result.userRequestModel.roles.map((e) => e.code).toList(),
+          "tenantId": envConfig.variables.tenantId,
+          "actionMaster": "actions-test",
+          "enabled": true,
+        },
+      );
       await localSecureStore.setBoundaryRefetch(true);
 
-      await localSecureStore.setRoleActions(actionsWrapper);
       emit(
         AuthAuthenticatedState(
           accessToken: result.accessToken,
@@ -103,7 +115,7 @@ class AuthBloc extends Bloc<AuthEvent, AuthState> {
       emit(const AuthUnauthenticatedState());
 
       AppLogger.instance.error(
-        title: 'Login error',
+        title: 'Erro de início de sessão',
         message: error.response?.data.toString(),
       );
     } catch (_) {
@@ -116,11 +128,33 @@ class AuthBloc extends Bloc<AuthEvent, AuthState> {
   //_onLogout event logs out the user and deletes the saved user details from local storage
   FutureOr<void> _onLogout(AuthLogoutEvent event, AuthEmitter emit) async {
     try {
-      emit(const AuthLoadingState());
-      await localSecureStore.deleteAll();
+      final isConnected = await getIsConnected();
+      if (isConnected) {
+        final accessToken = await localSecureStore.accessToken;
+        final user = await localSecureStore.userRequestModel;
+        final tenantId = user?.tenantId;
+        await authRepository.logOutUser(
+          logoutPath: Constants.logoutUserPath,
+          queryParameters: {
+            'tenantId': tenantId.toString(),
+          },
+          body: {'access_token': accessToken},
+        );
+        await localSecureStore.deleteAll();
+
+        emit(const AuthUnauthenticatedState());
+      }
     } catch (error) {
-      rethrow;
+      await localSecureStore.deleteAll();
+      emit(const AuthUnauthenticatedState());
     }
+  }
+
+  FutureOr<void> _onAuthLogoutWithoutToken(
+    AuthLogoutWithoutTokenEvent event,
+    AuthEmitter emit,
+  ) async {
+    await localSecureStore.deleteAll();
     emit(const AuthUnauthenticatedState());
   }
 }
@@ -138,6 +172,9 @@ class AuthEvent with _$AuthEvent {
   }) = AuthAutoLoginEvent;
 
   const factory AuthEvent.logout() = AuthLogoutEvent;
+
+  const factory AuthEvent.logoutWithoutAuthToken() =
+      AuthLogoutWithoutTokenEvent;
 }
 
 @freezed
